:white_check_mark: cleaned up test architecture.
