+++ src/rules/tests/rules/mod.rs
@@ -25,6 +25,65 @@ use crate::{
 };
 
 #[test]
+fn test_connect_rules() {
+    let re =
+        RuleEngine::new("./src/rules/tests/rules/connect").expect("couldn't build rule engine");
+    let mut state = get_default_state();
+
+    // ctx.client_addr is 0.0.0.0 by default.
+    state.get_context().write().unwrap().client_addr = "127.0.0.1:0".parse().unwrap();
+    assert_eq!(re.run_when(&mut state, "connect"), Status::Next);
+
+    state.get_context().write().unwrap().client_addr = "0.0.0.0:0".parse().unwrap();
+    assert_eq!(re.run_when(&mut state, "connect"), Status::Deny);
+}
+
+#[test]
+fn test_helo_rules() {
+    crate::receiver::test_helpers::logs::setup_logs();
+
+    let re = RuleEngine::new("./src/rules/tests/rules/helo").expect("couldn't build rule engine");
+
+    let mut state = get_default_state();
+    state.get_context().write().unwrap().envelop.helo = "viridit.com".to_string();
+
+    assert_eq!(re.run_when(&mut state, "connect"), Status::Next);
+    assert_eq!(re.run_when(&mut state, "helo"), Status::Next);
+}
+
+#[test]
+fn test_mail_from_rules() {
+    crate::receiver::test_helpers::logs::setup_logs();
+
+    let re = RuleEngine::new("./src/rules/tests/rules/mail").expect("couldn't build rule engine");
+
+    let mut state = get_default_state();
+    {
+        let email = state.get_context();
+        let mut email = email.write().unwrap();
+
+        email.envelop.mail_from = Address::new("staff@viridit.com").unwrap();
+        email.body = Body::Parsed(Box::new(
+            MailMimeParser::default()
+                .parse(
+                    br#"From: staff <staff@viridit.com>
+Date: Fri, 21 Nov 1997 10:01:10 -0600
+
+This is a reply to your hello."#,
+                )
+                .unwrap(),
+        ));
+    }
+
+    assert_eq!(re.run_when(&mut state, "mail"), Status::Accept);
+    assert_eq!(re.run_when(&mut state, "postq"), Status::Accept);
+    assert_eq!(
+        state.get_context().read().unwrap().envelop.mail_from.full(),
+        "no-reply@viridit.com"
+    );
+}
+
+#[test]
 fn test_rcpt_rules() {
     crate::receiver::test_helpers::logs::setup_logs();
